// src/app/models/Vehiculo.ts
export class Vehiculo {
    id_vehiculo?: number;
    id_conductor?: number;
    id_usuario?: number;
    placa?: string;
    color?: string;
    marca?: string;
    modelo?: string;
    estado?: number;
    descripcion?: string;
    sede?: string;
    impuestos?: number;
    año_vehiculo?: Date;
    año_compra?: Date;
    numero_chazis?: string;
    numero_motor?: string;
  }
  